support interactive to make things easier
